LPS-194418 Inline the component ObjectFilterParserServiceRegistry in ObjectEntryLocalServiceImp
