--- modules/apps/object/object-service/src/main/java/com/liferay/object/internal/filter/parser/ObjectFilterParserServiceRegistry.java
@@ -1,39 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.object.internal.filter.parser;
-
-import com.liferay.osgi.service.tracker.collections.map.ServiceTrackerMap;
-import com.liferay.osgi.service.tracker.collections.map.ServiceTrackerMapFactory;
-
-import org.osgi.framework.BundleContext;
-import org.osgi.service.component.annotations.Activate;
-import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Deactivate;
-
-/**
- * @author Feliphe Marinho
- */
-@Component(service = ObjectFilterParserServiceRegistry.class)
-public class ObjectFilterParserServiceRegistry {
-
-	public ObjectFilterParser getObjectFilterParser(String type) {
-		return _serviceTrackerMap.getService(type);
-	}
-
-	@Activate
-	protected void activate(BundleContext bundleContext) {
-		_serviceTrackerMap = ServiceTrackerMapFactory.openSingleValueMap(
-			bundleContext, ObjectFilterParser.class, "filter.type");
-	}
-
-	@Deactivate
-	protected void deactivate() {
-		_serviceTrackerMap.close();
-	}
-
-	private ServiceTrackerMap<String, ObjectFilterParser> _serviceTrackerMap;
-
-}
